requirements.txt
numba
numpy
Pillow
PySDL2
pysdl2-dll
